requirements.txt
brainflow
mido
mne
mne-realtime
numpy
PySimpleGUI
python-osc
pyserial
sentry-sdk
